\name{EcoVirtual-package}
\alias{EcoVirtual-package}
\alias{EcoVirtual}
\docType{package}
\title{
Simulating Metapopulation Dynamics
}
\description{
Simulating Ecological Models
}
\details{
\tabular{ll}{
Package: \tab EcoVirtual\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-02-08\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
metapop function
}
\author{
Alexandre Adalardo de Oliveira<aleadalardo@gmail.com>
}
\references{
Gotelli, N.J. 1991. Metapopulation models: the rescue effect, the propagule rain, and the core-satellite hypothesis. The American Naturalist 138:768-776.
}
\keyword{ package }
\examples{
metapop(tf=100,cl=10,ln=10,fi=0.5,pc=0.3,pe=0.15)
}
