\name{anima}
\alias{anima}
\alias{graf.fim}
\alias{meta.anima2}
\title{
EcoVirtual Graphics and Amimations
}
\description{
Graphics and animation of the simulation results
}
\usage{
anima(dados)
}
\arguments{
  \item{dados}{
Must be a 3D array producted by EcoVirtual function
}
}
\value{
Represent the simulation result in a graphic device
}
\author{
Alexandre Adalardo
}
\keyword{ simulation }
