\name{EcoVirtual-internal}
\title{Internal EcoVirtual objects}
\alias{MW}
\alias{MW.2.0}
\alias{anima.cena}
\alias{com.compete}
\alias{ddis.bc}
\alias{dnorm.trunc}
\alias{ext.game}
\alias{gr.toff}
\alias{graf.com}
\alias{grafeq}
\alias{iCol}
\alias{iRain}
\alias{ordena.polar}
\alias{pnorm.trunc}
\alias{prob.ssp}
\alias{reg.nicho}
\alias{rich}
\alias{sim}
\alias{sim.hub1}
\alias{sim.hub2}
\alias{sim.hub3}
\alias{spp.area}
\alias{sucMatrix}
\alias{compete}
\alias{iColExt}
\alias{rand.walk}
\description{Internal EcoVirtual objects.}
\details{These are not to be called by the user.}
\keyword{internal}
